Adding Basic Functions to the buttons
Restarted due to an error with deleting functions and trying to return to the visual winforms editor.
Creating some basic structure for the calculator. Creating some variables to hold inputs and output, as well as letting the number buttons add to the calculation. Also added functionality to the operator buttons, so that they could be added to the equations as well. The negate button will make a positive number negative and vice-versa. The clear buttons and the delete button work as intended as well.
Made some adjustments to the overal design of the calculator, making it easier to read.